requirements.txt
# arXiv Paper Bot - Python Dependencies

# Core dependencies
arxiv>=2.1.0                 # Official arXiv API client
PyYAML>=6.0                  # YAML configuration parsing
openai>=1.0.0                # OpenAI-compatible API client (for LLM summarization)
requests>=2.31.0            # HTTP 调用,用于消息推送

# Data handling
python-dateutil>=2.8.2       # Date parsing utilities
numpy>=1.24.0                # For numerical operations (used in personalization slots)

# LangGraph Agent Pipeline (for filter.mode: agent)
langgraph>=0.2.0             # LangGraph for agent orchestration
langchain-openai>=0.2.0      # OpenAI integration for LangChain
langchain-core>=0.3.0        # LangChain core components

# Optional: For scheduling
# schedule>=1.2.0            # Task scheduling (uncomment if needed)
# APScheduler>=3.10.0        # Advanced scheduling (alternative)

# Optional: Vector similarity features (not used in pure LLM mode)
# Uncomment if you want to add vector-based ranking:
# sentence-transformers>=2.0.0  # Scientific paper embeddings (SPECTER, SciBERT)
# chromadb>=0.4.0              # Vector database for similarity search
# faiss-cpu>=1.7.0             # Alternative: Facebook AI Similarity Search
# torch>=2.0.0                 # Required by sentence-transformers

# Development dependencies (optional)
# pytest>=7.4.0              # Testing framework
# black>=23.0.0              # Code formatting
# flake8>=6.1.0              # Linting
# mypy>=1.5.0                # Type checking
